pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Matvei Khmyzov <matw.hmizov@yandex.ru>"]
packages = [
    { include = "gendiff"}
]

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
makefile = "^1.1.0"
pytest = "^7.1.3"
flake8 = "^5.0.4"
coverage = "^6.5.0"
pytest-cov = "^4.0.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.generator_script:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
